corrects identification of Limited cards.
the templating is consistent throughout the card pool.
Case-sensitive "Limited." only appears on cards that are limited.
no need for regexp here.

while at it, i'm dropping the check on whether that flag has been set on the card
data directly since those data points will be removed in a follow-up
step.